requirements.txt
# Core FastAPI and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1

# Document processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
python-multipart==0.0.6

# Vector database and embeddings
faiss-cpu==1.7.4
pinecone-client==2.2.4
sentence-transformers==2.2.2
numpy==1.24.3

# LLM integration
openai==1.3.7

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
google-generativeai==0.3.1